requirements.txt
# Flask 核心
Flask==2.3.3
Flask-CORS==4.0.0

# Git 操作
GitPython==3.1.37

# Redis 缓存
redis==5.0.1

# HTTP 客户端
requests==2.31.0

# 环境变量
python-dotenv==1.0.0

# 生产环境服务器
gunicorn==21.2.0

# 开发和测试
pytest==7.4.3
pytest-cov==4.1.0

